Split out GeneralWorkflowIT.java tests
